+++ tests/commands/write_test.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from testing.files import FakeFileWriter
+from testing.files import FakeFileSystem
 from testing.logging import CapturedLogger
 from testing.rendering import FakeRenderer
 from write_config_files.commands.write import Writer
@@ -9,7 +9,7 @@ from write_config_files.config import TemplateFile
 
 
 def test_write_file():
-    file_writer = FakeFileWriter()
+    file_system = FakeFileSystem()
     logger = CapturedLogger()
 
     templates = TemplateConfig(
@@ -22,26 +22,26 @@ def test_write_file():
 
     writer = Writer(
         renderer=FakeRenderer('some content'),
-        file_writer=file_writer,
+        file_system=file_system,
         logger=logger,
     )
     writer.write(templates=templates, skip_if_exists=False)
 
-    assert file_writer.files == {
+    assert file_system.files == {
         'output/config-file': 'some content',
         'output/settings-file': 'some content',
     }
-    assert file_writer.executables == []
+    assert file_system.executables == []
 
     assert logger.captured == {
         'debug': [],
-        'info': [],
+        'info': ['writing output/config-file', 'writing output/settings-file'],
         'warn': [],
     }
 
 
 def test_write_executable_file():
-    file_writer = FakeFileWriter()
+    file_system = FakeFileSystem()
     logger = CapturedLogger()
 
     templates = TemplateConfig(
@@ -53,25 +53,25 @@ def test_write_executable_file():
 
     writer = Writer(
         renderer=FakeRenderer('some content'),
-        file_writer=file_writer,
+        file_system=file_system,
         logger=logger,
     )
     writer.write(templates=templates, skip_if_exists=False)
 
-    assert file_writer.files == {
+    assert file_system.files == {
         'output/a-script': 'some content',
     }
-    assert file_writer.executables == ['output/a-script']
+    assert file_system.executables == ['output/a-script']
 
     assert logger.captured == {
         'debug': [],
-        'info': [],
+        'info': ['writing output/a-script'],
         'warn': [],
     }
 
 
 def test_overwrite_existing_file():
-    file_writer = FakeFileWriter(files={'output/existing-file': ''})
+    file_system = FakeFileSystem(files={'output/existing-file': ''})
     logger = CapturedLogger()
 
     templates = TemplateConfig(
@@ -83,24 +83,24 @@ def test_overwrite_existing_file():
 
     writer = Writer(
         renderer=FakeRenderer('new content'),
-        file_writer=file_writer,
+        file_system=file_system,
         logger=logger,
     )
     writer.write(templates=templates, skip_if_exists=False)
 
-    assert file_writer.files == {
+    assert file_system.files == {
         'output/existing-file': 'new content',  # new content written
     }
 
     assert logger.captured == {
         'debug': [],
         'info': [],
-        'warn': [],
+        'warn': ['overwriting output/existing-file'],
     }
 
 
-def test_skip_existing_file():
-    file_writer = FakeFileWriter(files={'output/existing-file': ''})
+def test_skip_unchanged_file():
+    file_system = FakeFileSystem(files={'output/existing-file': 'some content'})
     logger = CapturedLogger()
 
     templates = TemplateConfig(
@@ -112,12 +112,41 @@ def test_skip_existing_file():
 
     writer = Writer(
         renderer=FakeRenderer('some content'),
-        file_writer=file_writer,
+        file_system=file_system,
+        logger=logger,
+    )
+    writer.write(templates=templates, skip_if_exists=False)
+
+    assert file_system.files == {
+        'output/existing-file': 'some content',  # same content
+    }
+
+    assert logger.captured == {
+        'debug': ['skipping output/existing-file because there are no changes'],
+        'info': [],
+        'warn': [],
+    }
+
+
+def test_skip_existing_file():
+    file_system = FakeFileSystem(files={'output/existing-file': ''})
+    logger = CapturedLogger()
+
+    templates = TemplateConfig(
+        root_dir='',
+        files=(
+            TemplateFile('templates/existing-file', 'output/existing-file', False),
+        ),
+    )
+
+    writer = Writer(
+        renderer=FakeRenderer('some content'),
+        file_system=file_system,
         logger=logger,
     )
     writer.write(templates=templates, skip_if_exists=True)
 
-    assert file_writer.files == {
+    assert file_system.files == {
         'output/existing-file': '',  # not changed
     }