+++ write_config_files/commands/diff.py
@@ -1,13 +1,17 @@
 from __future__ import annotations
 
 import difflib
+from typing import Protocol
 
 import attrs
 
 from ..config import TemplateConfig
-from ..files import Reader
 from ..printers import DiffPrinter
 from ..rendering import TemplateRenderer
+
+
+class Reader(Protocol):
+    def read_lines(self, file_path: str) -> list[str]: ...
 
 
 @attrs.frozen
@@ -18,7 +22,7 @@ class Differ:
 
     def print_diff(self, templates: TemplateConfig) -> None:
         for file in templates.files:
-            current_lines = self.reader.load_file(file.destination_path)
+            current_lines = self.reader.read_lines(file.destination_path)
 
             rendered_lines = (
                 self.renderer.render(file.template_name).splitlines(keepends=True)