+++ write_config_files/files.py
@@ -1,48 +1,30 @@
 from __future__ import annotations
 
 import os.path
-from typing import Protocol
 
 import attrs
 
-from .logging import Logger
-
-
-class Reader(Protocol):
-    def load_file(self, file_path: str) -> list[str]: ...
-
-
-class FileSystemReader:
-    def load_file(self, file_path: str) -> list[str]:
-        try:
-            with open(file_path) as f:
-                return f.readlines()
-        except FileNotFoundError:
-            return []
-
-
-class FileWriter(Protocol):
-    def file_exists(self, path: str) -> bool: ...
-    def write_file(self, path: str, content: str, is_executable: bool) -> None: ...
-
 
 @attrs.frozen
-class OsFileWriter:
-    logger: Logger
-
+class FileSystem:
     def file_exists(self, path: str) -> bool:
         return os.path.exists(path)
+
+    def read_content(self, path: str) -> str:
+        try:
+            with open(path) as f:
+                return f.read()
+        except FileNotFoundError:
+            return ''
+
+    def read_lines(self, path: str) -> list[str]:
+        return self.read_content(path).splitlines(keepends=True)
 
     def write_file(
             self,
             path: str, content: str,
             is_executable: bool,
     ) -> None:
-        if self.file_exists(path):
-            self.logger.warn(f'overwriting {path}')
-        else:
-            self.logger.info(f'writing {path}')
-
         os.makedirs(os.path.dirname(path), exist_ok=True)
         with open(path, 'w') as f:
             f.write(content)
@@ -51,15 +33,6 @@ class OsFileWriter:
             os.chmod(path, 0o744)
 
 
-@attrs.frozen
-class DryRunFileWriter:
-    logger: Logger
-
-    def file_exists(self, path: str) -> bool:
-        return os.path.exists(path)
-
+class DryRunFileWriter(FileSystem):
     def write_file(self, path: str, content: str, is_executable: bool) -> None:
-        if self.file_exists(path):
-            self.logger.warn(f'would overwrite {path}')
-        else:
-            self.logger.info(f'would write {path}')
+        pass